+++ README.md
@@ -3,15 +3,3 @@
 ## Overview
 
 MADSci is a modular, autonomous, and scalable framework for scientific discovery and experimentation.
-
-## Components
-
-- [Squid](./src/madsci.squid/README.md): The Workcell Management Engine.
-- [Types](./src/madsci.types/README.md): The Type Definition Library for MADSci.
-- [PyClient](./src/madsci.pyclient/README.md): The Python Client for MADSci.
-- [Module](./src/madsci.module/README.md): The Module Library for integrating devices.
-- [Server](./src/madsci.server/README.md): The REST API Server.
-- [CLI](./src/madsci.cli/README.md): The Command Line Interface.
-- [Dashboard](./src/madsci.dashboard/README.md): The web-based Dashboard and management interface.
-- [Resources](./src/madsci.resources/README.md): The Resource Library for managing resources.
-- [Events](./src/madsci.events/README.md): The Event Library for managing events.
