+++ madsci/madsci_node_module/madsci/node_module/rest_node_module.py
@@ -22,17 +22,15 @@ from madsci.common.types.action_types import ActionRequest, ActionResult, Action
 from madsci.common.types.admin_command_types import AdminCommandResponse
 from madsci.common.types.base_types import Error, new_ulid_str
 from madsci.common.types.event_types import Event
-from madsci.common.types.module_types import (
+from madsci.common.types.node_types import (
     AdminCommands,
     NodeCapabilities,
-)
-from madsci.common.types.node_types import (
     NodeInfo,
     NodeSetConfigResponse,
     NodeStatus,
 )
 from madsci.common.utils import threaded_task
-from madsci.module.abstract_module import (
+from madsci.node_module.abstract_node_module import (
     AbstractNode,
 )
