+++ pyproject.toml
@@ -9,7 +9,7 @@ dev = [
     "-e madsci.common @ file:///${PROJECT_ROOT}/madsci/madsci_common",
     "-e madsci.squid @ file:///${PROJECT_ROOT}/madsci/madsci_squid",
     "-e madsci.client @ file:///${PROJECT_ROOT}/madsci/madsci_client",
-    "-e madsci.module @ file:///${PROJECT_ROOT}/madsci/madsci_module",
+    "-e madsci.node_module @ file:///${PROJECT_ROOT}/madsci/madsci_node_module",
     "-e madsci.resource_manager @ file:///${PROJECT_ROOT}/madsci/madsci_resource_manager",
 ]
 [tool.pdm.version]
